refactor: Move Bundles components in one file
